solved some bugs; still kernel errors

+++ radis/gpu/structs.py
@@ -1,35 +1,35 @@
-from ctypes import Structure, c_float, c_int
+from ctypes import Structure, c_float, c_int, c_size_t
 
 
 class threadIdx_t(Structure):
     _fields_ = [
-        ("x", c_int),
-        ("y", c_int),
-        ("z", c_int),
+        ("x", c_size_t),
+        ("y", c_size_t),
+        ("z", c_size_t),
     ]
 
 
 class blockDim_t(Structure):
     _fields_ = [
-        ("x", c_int),
-        ("y", c_int),
-        ("z", c_int),
+        ("x", c_size_t),
+        ("y", c_size_t),
+        ("z", c_size_t),
     ]
 
 
 class blockIdx_t(Structure):
     _fields_ = [
-        ("x", c_int),
-        ("y", c_int),
-        ("z", c_int),
+        ("x", c_size_t),
+        ("y", c_size_t),
+        ("z", c_size_t),
     ]
 
 
 class gridDim_t(Structure):
     _fields_ = [
-        ("x", c_int),
-        ("y", c_int),
-        ("z", c_int),
+        ("x", c_size_t),
+        ("y", c_size_t),
+        ("z", c_size_t),
     ]
